Refactor LLM text generation native comps
Part work of code refactor to combine different text generation backends, remove duplcated native langchain and llama_index folder, consice the optimum habana implementation as a native integration OPEATextGen_Native.

Add feature for issue https://github.com/opea-project/GenAIComps/issues/998
