+++ .github/workflows/docker/compose/llms-compose.yaml
@@ -7,6 +7,10 @@ services:
     build:
       dockerfile: comps/llms/src/text-generation/Dockerfile
     image: ${REGISTRY:-opea}/llm-textgen:${TAG:-latest}
+  llm-textgen-gaudi:
+    build:
+      dockerfile: comps/llms/src/text-generation/Dockerfile.intel_hpu
+    image: ${REGISTRY:-opea}/llm-textgen-gaudi:${TAG:-latest}
   llm-ollama:
     build:
       dockerfile: comps/llms/text-generation/ollama/langchain/Dockerfile
@@ -19,14 +23,6 @@ services:
     build:
       dockerfile: comps/llms/src/faq-generation/Dockerfile
     image: ${REGISTRY:-opea}/llm-faqgen:${TAG:-latest}
-  llm-native:
-    build:
-      dockerfile: comps/llms/text-generation/native/langchain/Dockerfile
-    image: ${REGISTRY:-opea}/llm-native:${TAG:-latest}
-  llm-native-llamaindex:
-    build:
-      dockerfile: comps/llms/text-generation/native/llama_index/Dockerfile
-    image: ${REGISTRY:-opea}/llm-native-llamaindex:${TAG:-latest}
   llm-eval:
     build:
       dockerfile: comps/llms/utils/lm-eval/Dockerfile
