--- comps/llms/text-generation/native/langchain/requirements.txt
@@ -1,10 +1,0 @@
-docarray
-fastapi
-httpx
-langchain_core
-opentelemetry-api
-opentelemetry-exporter-otlp
-opentelemetry-sdk
-prometheus-fastapi-instrumentator
-shortuuid
-uvicorn
