--- comps/llms/text-generation/native/llama_index/requirements.txt
@@ -1,10 +1,0 @@
-docarray
-fastapi
-httpx
-llama_index
-opentelemetry-api
-opentelemetry-exporter-otlp
-opentelemetry-sdk
-prometheus-fastapi-instrumentator
-shortuuid
-uvicorn
