+++ tests/llms/test_llms_text-generation_native_on_intel_hpu.sh
@@ -12,34 +12,36 @@ function build_docker_images() {
     cd $WORKPATH
     docker build --no-cache \
         --build-arg https_proxy=$https_proxy --build-arg http_proxy=$http_proxy \
-        -t opea/llm-native:comps \
-        -f comps/llms/text-generation/native/langchain/Dockerfile .
+        -t opea/llm-textgen-gaudi:comps \
+        -f comps/llms/src/text-generation/Dockerfile.intel_hpu .
     if [ $? -ne 0 ]; then
-        echo "opea/llm-native built fail"
+        echo "opea/llm-textgen-gaudi built fail"
         exit 1
     else
-        echo "opea/llm-native built successful"
+        echo "opea/llm-textgen-gaudi built successful"
     fi
 }
 
 function start_service() {
-    LLM_NATIVE_MODEL="Qwen/Qwen2-7B-Instruct"
+    LLM_MODEL_ID="Qwen/Qwen2-7B-Instruct"
     llm_native_service_port=5070
     docker run -d \
-        --name="test-comps-llm-native-server" \
+        --name="test-comps-llm-textgen-gaudi-server" \
         -p ${llm_native_service_port}:9000 \
         --runtime=habana \
         --cap-add=SYS_NICE \
         --ipc=host \
+        -e LOGFLAG=True \
         -e http_proxy=${http_proxy} \
         -e https_proxy=${https_proxy} \
-        -e LLM_NATIVE_MODEL=${LLM_NATIVE_MODEL} \
+        -e LLM_MODEL_ID=${LLM_MODEL_ID} \
         -e HABANA_VISIBLE_DEVICES=all \
         -e OMPI_MCA_btl_vader_single_copy_mechanism=none \
         -e TOKENIZERS_PARALLELISM=false \
+        -e LLM_COMPONENT_NAME="OPEATextGen_Native" \
         --restart unless-stopped \
         --network bridge \
-        opea/llm-native:comps
+        opea/llm-textgen-gaudi:comps
 
     sleep 3m
 }
@@ -51,12 +53,12 @@ function validate_microservice() {
     HTTP_RESPONSE=$(curl --silent --write-out "HTTPSTATUS:%{http_code}" -X POST -d "$INPUT_DATA" -H 'Content-Type: application/json' "$URL")
     HTTP_STATUS=$(echo $HTTP_RESPONSE | tr -d '\n' | sed -e 's/.*HTTPSTATUS://')
     RESPONSE_BODY=$(echo $HTTP_RESPONSE | sed -e 's/HTTPSTATUS\:.*//g')
-    SERVICE_NAME="llm-native"
+    SERVICE_NAME="llm-textgen-gaudi"
 
     # check response status
     if [ "$HTTP_STATUS" -ne "200" ]; then
         echo "[ $SERVICE_NAME ] HTTP status is not 200. Received status was $HTTP_STATUS"
-        docker logs test-comps-llm-native-server >> ${LOG_PATH}/${SERVICE_NAME}.log
+        docker logs test-comps-llm-textgen-gaudi-server >> ${LOG_PATH}/${SERVICE_NAME}.log
         exit 1
     else
         echo "[ $SERVICE_NAME ] HTTP status is 200. Checking content..."
@@ -64,7 +66,7 @@ function validate_microservice() {
     # check response body
     if [[ "$RESPONSE_BODY" != *'"text":"What'* ]]; then
         echo "[ $SERVICE_NAME ] Content does not match the expected result: $RESPONSE_BODY"
-        docker logs test-comps-llm-native-server >> ${LOG_PATH}/${SERVICE_NAME}.log
+        docker logs test-comps-llm-textgen-gaudi-server >> ${LOG_PATH}/${SERVICE_NAME}.log
         exit 1
     else
         echo "[ $SERVICE_NAME ] Content is as expected."
@@ -72,7 +74,7 @@ function validate_microservice() {
 }
 
 function stop_docker() {
-    cid=$(docker ps -aq --filter "name=test-comps-llm-native*")
+    cid=$(docker ps -aq --filter "name=test-comps-llm-textgen-gaudi*")
     if [[ ! -z "$cid" ]]; then docker stop $cid && docker rm $cid && sleep 1s; fi
 }
